_archive/python_cli_backup/requirements.txt
typer>=0.9.0
rich>=13.4.2
pydantic>=2.0.0
loguru>=0.7.0
python-dotenv>=1.0.0

# GUI Automation
pyautogui==0.9.54
pillow==10.0.0
pytesseract==0.3.10
pyperclip==1.8.2
numpy==1.24.0
psutil==5.9.5
opencv-python==4.8.0.74  # For image recognition

# Platform-specific dependencies
# For Windows: pip install pywin32
# For macOS: No additional dependencies
# For Linux: sudo apt-get install scrot python3-tk python3-dev xdotool tesseract-ocr